final fixes (almost)

+++ client/src/appAuthenticated/pages/PageProfileEdit/BlockedList.js
@@ -1,4 +1,4 @@
-import React, { useContext } from 'react';
+import React, { useContext, useState } from 'react';
 import styled from 'styled-components';
 import axios from 'axios';
 import Separator from '../../../components/Separator';
@@ -6,6 +6,7 @@ import { faEyeSlash, faEye } from '@fortawesome/free-solid-svg-icons';
 import ProfileContext from '../../../contexts/ProfileContext';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { Image } from 'cloudinary-react';
+import { Redirect } from 'react-router-dom';
 
 
 const StyledSection = styled.section `
@@ -13,6 +14,7 @@ const StyledSection = styled.section `
     height:700px;
     border-radius:0 ${props => props.theme.borderRadius} ${props => props.theme.borderRadius} 0;
     background-color:#2b2c2e;
+    width:100%;
     @media (max-width: 1000px) { 
         border-radius:0;
     }
@@ -77,15 +79,20 @@ export default function BlockedList() {
 const authToken = localStorage.getItem('token');
 
     const profile = useContext(ProfileContext);
+    const [redirect, setRedirect] = useState(false);
     
     
     function User(props) {
-        const handleClick = event => {
-            const params = {data: {
-                type: "blocked",
-                targetUserId: props.user.userId,
-            }};
-            axios.delete(`/users/deleteRelationship?authToken=${authToken}`, params);
+        const handleClick = async event => {
+            const confirm = window.confirm(`Are you sure you want to unblock ${props.user.username} ?`)
+            if (confirm) {
+                const params = {data: {
+                    type: "blocked",
+                    targetUserId: props.user.userId,
+                }};
+                await axios.delete(`/users/deleteRelationship?authToken=${authToken}`, params);
+                setRedirect(true);
+            }
         }
         return (
             <StyledDiv>
@@ -94,7 +101,7 @@ const authToken = localStorage.getItem('token');
                     <Username>{props.user.username}</Username>
                     <Age>{props.user.age}, {props.user.city}</Age>
                 </InfosContainer>
-                <StyledButton onClick={handleClick} icon={faEye} size={"lg"}/>
+                <StyledButton onClick={() => handleClick()} icon={faEye} size={"lg"}/>
             </StyledDiv>
         )
     }
@@ -103,6 +110,7 @@ const authToken = localStorage.getItem('token');
         <StyledSection>
             <Separator icon={faEyeSlash} size={'lg'}/>
             {profile.blockedList.map(blockedUser => <User key={blockedUser.userId} user={blockedUser}>{blockedUser.username}</User>)}
+            {redirect && <Redirect to="/profile"/>}
         </StyledSection>
     )
 }
